requirements.txt
cffi==1.14.0
cryptography==2.9.2
Django==2.2.4
django-crispy-forms==1.7.2
Pillow==7.1.1
pkg-resources==0.0.0
pycparser==2.20
pytz==2019.2
six==1.14.0
sqlparse==0.3.0
django-import-export
psycopg2
